move JudgeClassifier to the ensemble module
